Update README.md
1 376 / 5 000
I am a professional with a solid career in the hotel and restaurant industry, with significant experience in hotel management and franchise development.

I have dual nationality, Portuguese and Angolan.

With an academic background that includes a degree in Hotel Tourism Business Management and a postgraduate degree in International Hotel Management, I am committed to providing high quality hospitality services.

My passion for food and wine led me to intern in the kitchen at DOM restaurant, by Chef Alex Atala, ranked 4th best in the world (in 2013), in São Paulo.
This allowed me to acquire valuable knowledge and develop a refined taste for cooking. I also have a course in Food Science with the Ferran Adrià seal of approval, which has increased my knowledge of the art of food.

I am known for my creativity and ability to transform challenges into opportunities.
I have a strong aptitude for communications, public relations and the ability to provide personalized customer experiences.
Furthermore, I try to stay up to date with the latest news in the sector, having completed several training sessions with Brazilian Tourism.

I am excited to continue my journey in the hospitality industry, contributing to the success of hotels, restaurants and franchises.